+++ exonum/src/blockchain/mod.rs
@@ -227,7 +227,6 @@ impl Blockchain {
             validator_keys: cfg.validator_keys,
             consensus: cfg.consensus,
             services: BTreeMap::new(),
-            majority_count: None,
         };
 
         let patch = {